Add support for the CMUX operation
Implement CMUX in Solidity as:
```
result = (ifTrue - ifFalse) * control + ifFalse
```

To implement, add the `Multiply` precompile in Solidity.

Add an example contract that shows how it works.

Also, add a Python test. Addresses, IPs and bytecode in the Python test
need to change. They are committed as is, because of impossibility to
run the Solidity compiler from Python on Linux on ARM. Fixing it is a
TODO for a future commit.