feat: use docker image to deploy fhevm SCs
This PR introduces:

- deployment of fhevm smart contracts through a docker image

The related file is docker-compose/docker-compose-fhevm-deploy.yml.
We need to copy as volume
- .env.example.deployment that includes freshly retrieved KMS signers addresses
- hardhat.config.ts to have a specific network (http://geth:8545)
- launch-fhevm-coprocessor.sh where we are building contracts and
  deploying all the fhEVM related SCs as ACL, GatewayContract,
TFHEExecutor...
- deploy-smart-contracts.sh which is the entrypoint and sequentially
  runnning the precompute addresse and launch-fhevm-coprocessor

TODO: add a specific network in hardhat.config.ts in the docker image

- running of test using main branch of  fhevm

We need to run some tests for decryption, reencryption and fhe
computation. They are (for now) defined in fhEVM repository.

We are cloning this repository and running some commands in
prepare-e2e-test target of Makefile to build artifacts and retrieve
original core contracts needed to run the hardhat tests.

TODO: all the tests files will be copied into this repository
(fhevm-devops). In this case we would not need to clone fhevm (main).