feat(input flow): complete input flow
Add a specific handler `input_handlers` that is responisble for sending
transaction and listen input response.

For now the full flow is not available, when we make a transaction to
Gateway to verify Input we listen the associated event (which is also listened
by the coprocessor) in order to consider the response.

Newt step could be to add thos mocked behavior in the mocked kms connector.

Also, in rollup, there is a match to distribute the corresponding event,
this behavior should change and handling the correct event should be done
directly in different handlers.

I am just wondering if we can register mutliple handler for one Relayer Event,
in this PR it is split for decryption flow `RelayerEventData::EventLogFromGwL2`
and input flow `InputEventData::EventLogFromGwL2`