feat(kms-connector): implement first version
This mock component is used to simulate kms connector.
As a recall, the kms connector role is to listen for input
and decryption request, process them and wait for associated
responses. Once response is received, a input/decrypt event
is emitted. This last event is catched back by the relayer
to process the response and return in to the user or to
http blockchain (public decyrpton only).

This component first version is integrated to existing orchestrator and
relayer events. It will be split in the next commit.

However this commit is still relevant to see how to move from an integrated
orchestrator with specific handler to a separate one.