chore: sync fhevm-devops with test cronjob
We used to apply a patch to ghcr.io/zama-ai/fhevm-e2e-test:v1.1.0 but
that isn't ideal, so this PR adds the patch to the actual repository so
that we can then create a new docker image with it.

It basically adds a reporter that outputs a json file.
This json file can then be used by a script to push metrics directly to
prometheus remote write endpoint.