feat(txn-sender): nonce management on errors
Since a nonce could be incremented in the `CachedNonceManager` when
transaction sending fails (say due to the GW being down or the
connection to it being lost), in this commit we introduce the
`NonceManagedProvider`. It contains a `CachedNonceManager` inside of it.
In case of any error, we re-create the `CachedNonceManager`, forcing it
to fetch the latest nonce from the GW and start from there. That allows
us to send txns concurrently in a safe way.

Add a configuration option for the required confirmations with a default
value of 0. Reasoning is that we assume the GW to be a rollup and we are
not waiting for hard finality on it.

We also simplify code by putting the DB pool and provider inside of the
the operations data structures.