feat: synchronize back instances (#230)
* feat: define sync service

* feat: sync between instances

It implements a syncrhonization solution using Redis queues.
Under the hood, it uses BullMQ to create a per-request queue and listens to jobs on
that queue.
It implements the synchronization for the validate address scenario.

* fix: broken tests

* feat: synch input proof

* fix: move syncronization event declaration into single use case

Instead of defining each event in the sync intance use case,
you now need to explicitly listen to event in each use case.

* fix: duplicated call due to wrong mapError implementation

* fix: broken tests

* docs: subscriptions & pubsub mechanisms (#244)

* docs: subscriptions & pubsub mechanisms

* fix: tapErr should be tapError

---------

Co-authored-by: Alessandro Vitali <alessandro.vitali@zama.ai>

* chore: fix makefile address validation event

---------

Co-authored-by: Gaspard Beernaert <gaspard@gmail.com>