feat(txn-sender): errors and retries improvements
Changes in this PR relate to adding ciphertext digests and ACL entries
to the GW as these txns are not supposed to fail and must eventually
succeed. Verify proofs are not subject to the same logic as they are
allowed to fail and the user must retry them manually.

Essentially, we handle transport errors such that they are retried
forever. We introduce a **transport** retry count that is different from
the retry count. The **transport** retry count can grow indefinitely.
We consider the following as transport errors:
 * TransportErrorKind::BackendGone - when the WS node is down
 * is_retry_err() - on HTTP status 429 and 503 and on
   MissingBatchResponse

Furthermore, log when max retry count is reached and the transaction is
not sent and is **not going** to be sent. Also, when a certain amount of
**transport** retries are reached, we also log (even though retries will
continue as it is a transport issue). These logs have the
`action = review` key-value, making it easy for operatins to identify
and do a manual review.

Reduce the retry count to 3 for verify proofs txns as these are allowed
to fail, forcing the user to retry them manually.

Other smaller improvements:
* use &[u8] instead of Vec<u8>
* set NOT NULL to fields that are not nullable
* put code in impl blocks where applicable
* add tests for stopping the Anvil node and recovering/retrying
* use match arms in error handling and not nested if/else inside arms