feat: add CachedNonceManagerWithRefresh
Added a draft for a method to cancel a transaction.
CachedNonceManager with sync method implemented.

Nonce is now auto-synced in the TransactionService if Nonce is too low.

Nonce too high isn't detected at the moment.
The only way to handle it is to have a timeout based on the slot-time and
number of slots passed since the tx was emitted to the mem-pool.
This is still work to be done.
(could be added in the maintenance task of the TransactionService)

Remove request-id:on-chain-id 1:1 mapping assumption in input handler.

Gracefully shutdown all tokio tasks. Some works still needs to be done
on that to properly finish all tasks. I'm not sure the complexity needed
to do so properly is worth the effort at the moment.