feat(coprocessor): add log level configurations
Add the `log-level` cmd line option to configure the log level.

Also, make sure gw-listener and transaction-sender retry connecting to
the node infinitely at startup as the `WsConnect` options don't apply
to connect during provider creation.

Finally, add structured logging to fhevm-listener instead of println!s.
Also, added some TODOs to remove panics and, instead, propagate errors.
Will be done in a separate PR.

As a side note, using structured logging in all other services where it
is applicable will be done in a separate PR as well (as we usually only
utilize the `message` field in the log macros).