feat: integrate fhevm-relayer (#319)
https://github.com/zama-ai/fhevm-relayer/pull/339 implements the proper
SQS interactions to integrate the fhevm-relayer with the Console.
`make fhevm-up && make console-build && FHEVM_RELAYER=1 make console-up && make test`
can be used to test that everything works fine.

Tests are passing when interacting with the Console through SQS.
The serialization format of messages is inconsistent between operations.
It should be refactored at a later time.

`payload` should contain the raw http payload.
`requestId` should be at the same level as `payload` and not inside it.
User decrypt payload should be consistent with the others.
Public decrypt SQS endpoint is not yet tested.