feat(coprocessor): add context ID for input verification
Handle the `context_id` field in the `VerifyProofRequest` event,
transport it to the DB and use it to sign the EIP-712 message in the
transaction-sender when sending the `verifyProofResponse` transaction.

Replaces the existing https://github.com/zama-ai/fhevm/pull/545 PR,
thanks to @RomanBredehoft for the initial work.

Based on https://github.com/zama-ai/fhevm/pull/211