feat: add user-decrypt cache (#355)
This commit adds a user-decryption cache for requests and responses.
The requests cache stores the following mappings:
- request-cache: hash(request) -> decryption-id
- response-cache: decryption-id -> response (edited)

Each cache storage is tested and a specific test was added to the "end-to-end" tests to
make sure that multiple requests with the same payload are always met in under a second,
signifying that the cache was indeed hit.

Anything that is meant to be persistent shouldn't use the internal
request-id.

On a side note I modified a bit the tests that leverage docker compose to add a flag to not
shutdown the containers to make repetitive testing faster.
Towards that goal I also set a specific target directory for the build used in the test to
avoid any conflict with other builds that might occur on the local machine.

I also slightly modified the tests to remove the database of the relayer
to avoid conflicts between tests.
I also modified the logging settings to respect the `NO_COLOR` env var.
https://no-color.org/ for more context.