* feat: user-decrypt cache improvements, pub-dec cascade + tx metrics
* feat: user-decrypt cache improvements

This commit follows the modifications done by Mano in the
public-decryption cache to handle in-fly requests (i.e. requests that
are waiting for a transaction request to be emitted).

It also modifies the way the key of user-decrypt request is generated to
not include the EIP-712 signature and request_validity.
This avoids having cache-miss when a dapp changes the EIP-712.

When the first call to get_value of a user-decrypt-request cache is
done a lock on said key is applied to avoid multiple calls until the
leader request either succeeds in persisiting a value, or the lock is
unlocked due to a failure.

This is done by having an infinite loop of cascading leaders.

Also this commit add transaction metrics.

* chore: add cascading leaders to pub-dec cache