fix(coprocessor): do not expand  an empty compact ct list
Current version of tfhe-rs panics when trying to expand and verify an
empty compact ciphertext list. We add a check to avoid this if the list
is empty.

We consider an empty list to be a valid empty input (with the ZK proof
verified in the future when tfhe-rs is updated).

Future versions of tfhe-rs may handle this gracefully, in which case we
can remove this check.

Add a test to verify that an empty compact ciphertext list does not
cause a panic.