fix(oracle): add same gateway decryption contract bit limitation in oracle [API-695] (#564)
The decryption contract on the gateway chain has a constrain on the
number of bits per public decryption request.
The easiest way to avoid a mismatch between the host and the gateway is
to check for this limitation directly in the oracle decryption contract.

That being said we should think of another solution in the future.
Indeed a contract might need to do an oracle call with a number of bits
higher than the current limit.
The relayer should be able to split and gather all responses but in the
current design in would cause an issue when verifying the signature of
the decryption results.