feat(coprocessor): handle graceful shutdown cases in txn-sender
This commit assumes that `provider_max_retries` and
`provider_retry_interval` are set such that it would take a long time
until the provider stops retrying. If that happens, we consider the
provider no longer usable and we shut down the txn-sender. This keeps
code simpler and avoids the need to refactor and retest retry behaviour.
Reason we do it is that once retries are exhausted, the provider is no
longer usable and `BackendGone` is returned.

We could consider using HTTP instead of WS, but that is left for future
investigation.

The following other changes are made:
 * propagete errors without converting them to strings, e.g. `bail(a)`
   instead of `bail("xxx {}", e)`
 * remove the `provider()` and `check_provider_connection()` functions
   from the `TransactionOperation` trait as they are used for health
   checking and we do not need the health check to go to just the first
   operation - instead, we use the provider from the `TransactionSender`
* since the provider can retry for a long time, if a signal to stop the
  sender is received, we add a graceful timeout such that the sender
  stops in a timely manner and doesn't wait for a long time for the
  provider - for that, we introduce the `graceful_shutdown_timeout`
  config option
* plug the cancellation token to `get_chain_id()` on startup, such that
  the sender can stop in a timely manner
* use `tokio::spawn()` instead of join! for starting the sender and the
  HTTP server