feat(coprocessor): send sync and nonce mngmt fixes in txn-sender (#1388)
* Use eth_sendRawTransactionSync for sending txns to the GW
* Do not consume a nonce if gas estimation fails due to a revert
* Default error-sleep-max-secs from 300 to 4 seconds
* Deprecate required-txn-confirmations config param
* Use txn-receipt-timeout-secs conf param as send sync timeout and set
default from 10 to 4 seconds
* Update alloy version to 1.1.2 for Provider::send_transaction_sync
* Deprecate retry-immediately-on-nonce-error

Above should fix issues that were leading to constant retries due to a
txn that gets reverted, leading to a consumed nonce. Furthermore, using
send sync should decrease latency and load.

For now, we keep the strategy of redoing the whole batch on failure. We
might change it in a future commit in order to keep this one simpler.

Cherry-pick from https://github.com/zama-ai/fhevm/pull/1388